requirements.txt
discord.py
llama-cpp-python